ListNetworked windows

+++ xrect_cut.h
@@ -4,28 +4,28 @@
 static Rect2
 xcut_left(Rect2 *rect, f32 a) {
     f32 minx = rect->min[0];
-    rect->min[0] = fmin(rect->max[0], rect->min[0] + a);
+    rect->min[0] = (f32)fmin(rect->max[0], rect->min[0] + a);
     return (Rect2){ minx, rect->min[1], rect->min[0], rect->max[1] };
 }
 
 static Rect2
 xcut_right(Rect2 *rect, f32 a) {
     f32 maxx = rect->max[0];
-    rect->max[0] = fmax(rect->min[0], rect->max[0] - a);
+    rect->max[0] = (f32)fmax(rect->min[0], rect->max[0] - a);
     return (Rect2){ rect->max[0], rect->min[1], maxx, rect->max[1] };
 }
 
 static Rect2
 xcut_top(Rect2 *rect, f32 a) {
     f32 miny = rect->min[1];
-    rect->min[1] = fmin(rect->max[1], rect->min[1] + a);
+    rect->min[1] = (f32)fmin(rect->max[1], rect->min[1] + a);
     return (Rect2){ rect->min[0], miny, rect->max[0], rect->min[1] };
 }
 
 static Rect2
 xcut_bottom(Rect2 *rect, f32 a) {
     f32 maxy = rect->max[1];
-    rect->max[1] = fmax(rect->min[1], rect->max[1] - a);
+    rect->max[1] = (f32)fmax(rect->min[1], rect->max[1] - a);
     return (Rect2){ rect->min[1], rect->max[1], rect->max[0], maxy };
 }
